Mobile_flask_app/requirements.txt
azure-common==1.1.25
azure-iot-device==2.1.4
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
beautifulsoup4==4.9.1
Bottleneck==1.2.1
bs4==0.0.1
bson==0.5.8
certifi==2020.4.5.1
cffi==1.12.3
chardet==3.0.4
chardet==3.0.4
Click==7.0
cloudant==2.13.0
cloudpickle==1.2.1
colorama==0.4.1
comtypes==1.1.7
contextlib2==0.5.5
dask==2.1.0
dataclasses==0.6
decorator==4.4.0
defusedxml==0.6.0
distlib==0.3.0
distributed==2.1.0
docopt==0.6.2
docutils==0.14
entrypoints==0.3
environment-kernels==1.1.1
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.2
Flask-RESTful==0.3.8
fsspec==0.7.1
future==0.17.1
gast==0.3.3
glob2==0.7
greenlet==0.4.15
grpcio==1.29.0
heapdict==1.0.0
idna==2.10
imageio==2.5.0
imagesize==1.1.0
janus==0.4.0
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.13.2
json5==0.8.4
jsonschema==3.0.1
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
libarchive-c==2.8
numba==0.44.1
numexpr==2.6.9
numpy==1.19.0
olefile==0.46
openpyxl==2.6.2
packaging==19.0
paho-mqtt==1.5.0
pandas==1.0.5
pandocfilters==1.4.2
parso==0.5.0
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.4
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
pigar==0.10.0
pipreqs==0.4.10
pkginfo==1.5.0.1
pluggy==0.12.0
ply==3.11
protobuf==3.12.1
psutil==5.6.3
py==1.8.0
pyodbc==4.0.26
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.0
random2==1.0.1
ratelim==0.1.6
requests==2.24.0
requests-unixsocket==0.2.0
rope==0.14.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
soupsieve==2.0.1
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.25.9
wcwidth==0.1.7
Werkzeug==1.0.1
wrapt==1.11.2
scipy==1.4.1
scikit-image==0.17.2
matplotlib==3.2.2
scikit-learn==0.20.0
Keras==2.1.5
tensorflow==2.2.0rc1
gevent==20.6.2










